requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
gradio
fastapi
uvicorn
httpx
pytest
python-dotenv
requests
huggingface_hub
